Rename DatasetPipelines to Datasets.Pipelines
Renames src/EncDotNet.S100.DatasetPipelines to
src/EncDotNet.S100.Datasets.Pipelines so the project name slots
naturally alongside the per-spec dataset libraries
(EncDotNet.S100.Datasets.S101, .S102, etc.) under the shared
"EncDotNet.S100.Datasets.*" prefix.

- Renames the directory, csproj, and root namespace.
- Fully qualifies references to EncDotNet.S100.Pipelines.Vector.Feature
  in S101DatasetProcessor.cs since the new namespace shadows the
  framework Pipelines namespace via name resolution.
- Updates the Viewer csproj reference, viewer using directives, the
  visual regression library, and EncDotNet.S100.slnx.
