+++ EncDotNet.S100.slnx
@@ -8,7 +8,7 @@
     <Project Path="src/EncDotNet.S100.Datasets.S124/EncDotNet.S100.Datasets.S124.csproj" />
     <Project Path="src/EncDotNet.S100.Datasets.S129/EncDotNet.S100.Datasets.S129.csproj" />
     <Project Path="src/EncDotNet.S100.Datasets.S421/EncDotNet.S100.Datasets.S421.csproj" />
-    <Project Path="src/EncDotNet.S100.DatasetPipelines/EncDotNet.S100.DatasetPipelines.csproj" />
+    <Project Path="src/EncDotNet.S100.Datasets.Pipelines/EncDotNet.S100.Datasets.Pipelines.csproj" />
     <Project Path="src/EncDotNet.S100.ExchangeSets/EncDotNet.S100.ExchangeSets.csproj" />
     <Project Path="src/EncDotNet.S100.Features/EncDotNet.S100.Features.csproj" />
     <Project Path="src/EncDotNet.S100.Hdf5.PureHdf/EncDotNet.S100.Hdf5.PureHdf.csproj" />
