+++ src/EncDotNet.S100.Viewer/Views/DatasetsView.axaml.cs
@@ -2,7 +2,7 @@ using Avalonia.Controls;
 using Avalonia.Input;
 using Avalonia.Interactivity;
 using Avalonia.Platform.Storage;
-using EncDotNet.S100.DatasetPipelines;
+using EncDotNet.S100.Datasets.Pipelines;
 using EncDotNet.S100.Viewer.ViewModels;
 
 namespace EncDotNet.S100.Viewer.Views;
